ajaxcontent: parse 'lecturers' array
This commit joins the defined lecturers Array to a human-readable String.

eLectures will slightly change the returned JSON schema for the
public_schedule service. The system allows for multiple lecturers per
lecture, so the returned data should reflect that properly.

Upstream will soon(tm) respond with the `lecturers` attribute, which is an
array of Strings with 0 or more members. The previous `lecturer` attribute
will be omitted.