Fix #14 (switch to different mensaplan XML feed)
The Studiwerk discontinued the old feed, but thanks to the active OpenMensa community there's now a parser for the new source. Its output is very similar, so just a few changes were necessary. For a nicer output, I also added the term "mit drei Beilagen" to the search&replace list.