Fix regenradar by switching to new wetteronline.de API
I tried to document as much as reasonably possible directly in the code and comments. More information on wetteronline.de's new API is gathered here: https://github.com/christophfriedrich/wetteronline-regenradar

Note that the bounding box visible in the GIF is a lot smaller than before (previously it was Lingen to Düsseldorf and Apeldoorn to Warburg, now it's Rheine to Duisburg and Apeldoorn to Arnsberg).