Rewrite Mensaplan to reflect change of price policy and data structure
Meals are now no longer priced per menu, but component-wise, making it interesting to also output the side dishes and their prices.

Also, it's no longer the case that each meal is put in its own meal category (which didn't make sense anyway), but some categories have multiple meals. Most notably this is the case for the three main components of the upper area, which is the reason why some of them were missing in the previous output.

With these changes, it makes more sense to output the offer as multiple lists with subheadings, rather than a table, so I changed the markup and adapted the CSS accordingly.